requirements.txt
fastapi[all]
sqlalchemy 
mysql-connector-python
passlib[argon2]
pyjwt
python-jose